% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_save_footnote.R
\name{format_save_footnote}
\alias{format_save_footnote}
\title{Save a LaTeX-formatted string as a .tex document to be included as a table footnote}
\usage{
format_save_footnote(text = "", filename, stars = FALSE)
}
\arguments{
\item{text}{LaTeX-formatted string.}

\item{filename}{File name of corresponding graph or table.}

\item{stars}{Indicates whether explanation for significance stars should be included or not. Default is FALSE.}
}
\description{
Save a LaTeX-formatted string as a .tex document to be included as a table footnote
}
\examples{
library(stringr)

# Define footnote
footnote <- paste0("This is a footnote. \n",
                   "It has two rows and includes significance stars.")

# Save footnote with significance stars
format_save_footnote(footnote, "current_table.tex", TRUE)

# In LaTeX document:
# \input{../results/tables/current_table.tex}
# \input{../results/notes/current_table.tex}
}
